Edited iBeacon Distance Compare Methodthology

+++ app/src/main/java/com/melody/mobile/android/testibeacon/ibeacon/IBeacon.java
@@ -66,7 +66,7 @@ public class IBeacon implements Serializable{
 	/**
 	 * A calculated proximity of the iBeacon based on <code>_powerValue</code>
 	 */	
-	private int _proximity;
+	private double _proximity;
 
 	/**
 	 * The MAC address reported by the iBeacon
@@ -111,7 +111,7 @@ public class IBeacon implements Serializable{
 	 * @param minor The Minor number (1-65535)
 	 * @param proximity Proximity in meters
 	 */
-	public IBeacon(byte[] uuid, int major, int minor, int proximity) {
+	public IBeacon(byte[] uuid, int major, int minor, double proximity) {
 
 		_uuid = uuid;
 		_major = major;
@@ -236,14 +236,13 @@ public class IBeacon implements Serializable{
 		this._txPower = _txPower;
 	}
 
-	public int getProximity() {
+	public double getProximity() {
 		return _proximity;
 	}
 	
-	public void setProximity(int _proximity) {
+	public void setProximity(double _proximity) {
 		this._proximity = _proximity;
 	}
-	
 
 	public int getBattery() {
 		return _battery;
@@ -283,7 +282,8 @@ public class IBeacon implements Serializable{
 	 */
 	@Override
 	public boolean equals(Object obj) {
-	    if (obj == null) {
+
+		if (obj == null) {
 	        return false;
 	    }
 	    if (getClass() != obj.getClass()) {
@@ -317,5 +317,4 @@ public class IBeacon implements Serializable{
 
 		return "UUID:" + this.getUuidHexString() + " M:" + this.getMajor() + " m:" + this.getMinor() + " p:" + this.getProximity();
 	}
-
 }
